rename component, add grid, fix styles for reservations section
