+++ src/Reservations/index.js
@@ -1,11 +1,11 @@
 import React from "react";
-import styled from "styled-components";
+import { Grid, Row, Col } from "react-flexbox-grid";
 
-import { Container, CardGroupRow, Row, Column, Title } from "../commons/styled";
+import { Slider, Slide, SectionTitle } from "../commons/styled";
 import SeeAllLink from "../commons/seeAllLink";
 import NextLink from "../commons/nextLink";
 
-import Card from "./cardComponent";
+import Card from "./Card";
 
 import chumley from "./chumley.png";
 import hanjan from "./hanjan.png";
@@ -14,54 +14,56 @@ import seaprice from "./seaprice.png";
 
 export default function() {
   return (
-    <Container>
-      <Row justify="flex-start" align="center">
-        <Column size={11}>
-          <Title>Popular reservations around the world</Title>
-        </Column>
-        <Column>
-          <SeeAllLink />
-        </Column>
-      </Row>
-      <CardGroupRow>
-        <Column size={3}>
-          <Card
-            title="Chumley's"
-            img={chumley}
-            alt="Chumley's"
-            category="speakeasy"
-            price={60}
-          />
-        </Column>
-        <Column size={3}>
-          <Card
-            title="Hanjan"
-            img={hanjan}
-            alt="Hanjan"
-            category="korean gastropub"
-            price={50}
-          />
-        </Column>
-        <Column size={3}>
-          <Card
-            title="Hanjan"
-            img={meats}
-            alt="Prime Meats"
-            category="german american"
-            price={55}
-          />
-        </Column>
-        <Column size={3}>
-          <Card
-            title="Seaprice"
-            img={seaprice}
-            alt="Seaprice"
-            category="fine seafood"
-            price={70}
-          />
-        </Column>
-        <NextLink top={72} />
-      </CardGroupRow>
-    </Container>
+    <section>
+      <Grid>
+        <Row middle="xs">
+          <Col xs={10} md={10} lg={8}>
+            <SectionTitle>Popular reservations around the world</SectionTitle>
+          </Col>
+          <Col xs={2} md={2} lg={1} lgOffset={3}>
+            <SeeAllLink />
+          </Col>
+        </Row>
+        <Slider>
+          <Slide xs={6} md={4} lg={3}>
+            <Card
+              title="Chumley's"
+              img={chumley}
+              alt="Chumley's"
+              category="speakeasy"
+              price={60}
+            />
+          </Slide>
+          <Slide xs={6} md={4} lg={3}>
+            <Card
+              title="Hanjan"
+              img={hanjan}
+              alt="Hanjan"
+              category="korean gastropub"
+              price={50}
+            />
+          </Slide>
+          <Slide xs={6} md={4} lg={3}>
+            <Card
+              title="Hanjan"
+              img={meats}
+              alt="Prime Meats"
+              category="german american"
+              price={55}
+            />
+          </Slide>
+          <Slide xs={6} md={4} lg={3}>
+            <Card
+              title="Seaprice"
+              img={seaprice}
+              alt="Seaprice"
+              category="fine seafood"
+              price={70}
+            />
+          </Slide>
+          <NextLink top={30} />
+        </Slider>
+      </Grid>
+    </section>
   );
 }
